pyproject.toml
[tool.poetry]
name = "fastapi-structure"
version = "0.1.0"
description = "A structured FastAPI project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
kafka-python = "^2.0.2"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.0.6"
pydantic = "^2.7.4"
alembic = "^1.13.1"
celery-flower = "^2.0.1"
greenlet = "^3.0.1"
gevent = "^24.2.1"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
